kill ADB only on unload instead of every time the Dashboard closes
Ideally, we'd continue killing ADB every time the Dashboard closes.
But a new ADB service instance can't reconnect to a device to which
another instance previously connected until the device is replugged
into the computer or its adbd daemon is restarted.  So we're better
off leaving ADB running while the addon is loaded.