Basic 'new project' support.
Added a "New Project" button to the UI, that asks the user for an app name, then creates a copy of an empty app template at the location of the user's choosing, placing basic files needed to have a working application. Next to come is multiple template support, and the ability to modify more properties of the manifest itself. New in this commit is the DevKit module, which is the home for this functionality in the event we fragment the Simulator in future releases.