Better error message if ADB is already running
Detect if it is running based on failiure to bind 5037. The tests still
correctly identify if an actual adb background process is runnning even
though it binds to 5037. However, if an adb background process is not
running and 5037 is in use, chances are that the Simulator was started
in another profile.